requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
anthropic==0.86.0
llama-index==0.14.18
llama-index-embeddings-fastembed==0.6.0
llama-index-llms-anthropic==0.11.1
fastembed==0.7.4
youtube-transcript-api==0.6.3
trafilatura==1.12.2
